Fix a type issue and refactor the type check tests
